Add JSON input normalization
TypeForm sends incoming data in UTF8 including spaces in multiple formats.
This presents a challenge when trying to match against text in the VirtualModel
implementation. To address this, we now remove both charpoint 160
(non-breaking space) and \\u00a0, to allow regular expressions to match
the text structure of questions.

This adds a #raw_json accessor to both form and webhook classes, to
allow access to the original JSON, while relying on the normalized JSON
for behavior & data transformations.