Strip hex-formatted non-breaking UTF-8 spaces
Typeform recently began sending an additional representation
of breaking space (hex-formatted). This adds it to the list
of characters in the space normalizing regex, replacing
it with a normal space.

Adds \xC2\xA0 in the single form example JSON, to ensure
that normalization is working on both sets of possible spaces.